Fix QtakesServiceTest's annotations
The @BeforeClass and @After annotations are JUnit 4 annotations, and
shouldn't be used in JUnit Jupiter tests.
Moreover, there is a subtle bug here - the mock server is started once
for the entire test class (with the @BeforeClass annotation), but
stopped after every test (with the @After annotation). This goes
unnoticed since the class contains only one test, but it is still
wrong, and will prevent adding more tests to this class in the future.

This patch solves both issues and replaces the annotations with JUnit
Jupiter's @BeforeAll and @AfterAll, respectively.